routes user

+++ server/routes/item.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const Item = require('../models/Item');
 const { route } = require('./auth');
+const requireAuth = require('../middlewares/requireAuth')
 
 // GET	/api/items	Gets all the items in the DB
 router.get('/', (req, res) => {
@@ -22,8 +23,11 @@ router.get('/:id', (req, res) => {
 });
 
 // POST	/api/items	Create an item in the DB	Requires auth.
-router.post('/', (req, res) => {
-  Item.create(req.body)
+router.post('/', requireAuth, (req, res) => {
+  const item = req.body;
+  item.creator = req.session.currentUser
+
+  Item.create(item)
     .then((dbRes) => {
       res.status(201).json(dbRes);
     })
@@ -31,22 +35,38 @@ router.post('/', (req, res) => {
 });
 
 // PATCH	/api/items/:id	Update an item	Requires auth.
-router.patch('/:id', async (req, res) => {
+router.patch('/:id', requireAuth, async (req, res) => {
   try {
-    const dbRes = await Item.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-    });
-    res.status(200).json(dbRes);
+    const foundItem = await Item.findById(req.params.id)
+    const currentUser = req.session.currentUser.toString()
+
+    if(foundItem.creator.toString() === currentUser) {
+      const dbRes = await Item.findByIdAndUpdate(req.params.id, foundItem, {
+        new: true,
+      });
+      res.status(200).json(dbRes);
+    } else {
+      res.status(403).json("you shall not pass")
+    }
   } catch (err) {
     res.status(500).json(err);
   }
 });
 
 // DELETE	/api/items/:id	Deletes an item	Requires auth.
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', requireAuth, async (req, res) => {
   try {
-    await Item.findByIdAndDelete(req.params.id);
-    res.status(200).json({ message: 'Successfully deleted.' });
+
+    const foundItem = await Item.findById(req.params.id)
+    const currentUser = req.session.currentUser.toString()
+
+    if(foundItem.creator.toString() === currentUser) {
+      await Item.findByIdAndDelete(req.params.id);
+      res.status(200).json({ message: 'Successfully deleted.' });
+    } else {
+      res.status(403).json("you shall not pass")
+    }
+    
   } catch (err) {
     res.status(500).json(err);
   }
